refactor ♻️: apply review fixes to ai and git helpers (#39)
## Summary

- isolate AI communication into new `client.py`.
- move commit history helpers into dedicated `history.py`.
- document new structure and refactor details.
