+++ git_acp/git/__init__.py
@@ -3,18 +3,20 @@
 from git_acp.git.classification import CommitType, classify_commit_type
 from git_acp.git.git_operations import (
     GitError,
-    analyze_commit_patterns,
-    find_related_commits,
     get_changed_files,
     get_current_branch,
     get_diff,
-    get_recent_commits,
     git_add,
     git_commit,
     git_push,
     run_git_command,
     setup_signal_handlers,
     unstage_files,
+)
+from git_acp.git.history import (
+    analyze_commit_patterns,
+    find_related_commits,
+    get_recent_commits,
 )
 
 __all__ = [
